making responsive
